开始生成mips代码

+++ C0_mips_vscopy/C0_mips/C0_mips/middlecode.cpp
@@ -10,7 +10,7 @@ void ConstMidCode(string op, string type, int num, string name)
 	code.op = op;
 	code.ob1 = type;
 	stringstream s_num;
-	s_num << num;
+	s_num << (int)num;
 	code.ob2 = s_num.str();
 	code.ob3 = name;
 	middlecode_list.push_back(code);
@@ -25,7 +25,7 @@ void VarMidCode(string type, int length, string name)
 	if (length != 0)
 	{
 		stringstream s_length;
-		s_length << length;
+		s_length << (int)length;
 		code.ob2 = s_length.str();
 	}
 	else
